Added Inbounds, insights, inventory, Invoices and reductions support & Testing for every implemented function

+++ Twinvision.BolRetailerApi/ObjectDefinitions/Inbound.cs
@@ -6,15 +6,15 @@ namespace Twinvision.BolRetailerApi
 {
     public class Inbound
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Reference { get; set; }
         public DateTime CreationDate { get; set; }
         public string State { get; set; }
         public bool LabellingService { get; set; }
-        public int AnnouncedBSKUs { get; set; }
-        public int AnnouncedQuantity { get; set; }
-        public int ReceivedBSKUs { get; set; }
-        public int ReceivedQuantity { get; set; }
+        public long AnnouncedBSKUs { get; set; }
+        public long AnnouncedQuantity { get; set; }
+        public long ReceivedBSKUs { get; set; }
+        public long ReceivedQuantity { get; set; }
         public Timeslot TimeSlot { get; set; }
         public Product[] Products { get; set; }
         public StateTransition[] StateTransitions { get; set; }
